Console-Based Quiz Application
The project is a console-based quiz application developed using core Java concepts. It features a set of 10 questions, each with 4 options, where one option is the correct answer. The application provides an interactive user experience by including lifeline options such as 50-50 and Audience Poll, as well as an option to quit the game at any point. The lifelines are designed to assist the user in making informed decisions, while the scoring system rewards 100 points for each correct answer. Incorrect answers result in the termination of the game.